Abstract models to separate files
