+++ app/models/__init__.py
@@ -1,0 +1,3 @@
+from .user import User
+from .market import Market
+from .favorite import Favorite
