+++ app/models/user.py
@@ -1,0 +1,14 @@
+from sqlalchemy import Integer, String, Column
+from sqlalchemy.orm import relationship
+from app.database import Base
+
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True, index=True)
+    email = Column(String, unique=True, index=True)
+    hashed_password = Column(String, nullable=False)
+    favorites = relationship(
+        "Market",
+        secondary="favorites",
+        backref="users"
+    )
